frontend/requirements.txt
gradio==4.29.0
pillow==10.2.0
python_dotenv==1.0.1